backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
google-genai>=1.0.0
aiosmtplib>=2.0.0
markdown>=3.5.0
httpx>=0.25.0
openpyxl>=3.1.0
pandas>=2.0.0
pytrends>=4.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0
